Add admin blueprint
