--- application/templates/admin.html
@@ -1,23 +1,0 @@
-{% extends "layout.html" %}
-
-{% block title %}
-    Admin Panel
-{% endblock %}
-
-{% block main %}
-    <div class="container inner">
-        <h1>Howdy!</h1>
-        {% with messages = get_flashed_messages() %}
-            {% if messages %}
-                {% for message in messages %}
-                    <h3>{{ message }}</h3>
-                {% endfor %}
-            {% endif %}
-        {% endwith %}
-        <a class="btn_a" href='/create'><button class="btn admin-btn">Create Post</button></a>
-        <a class="btn_a" href='/edit'><button class="btn admin-btn">Edit Post</button></a>
-        <a class="btn_a" href='/delete'><button class="btn admin-btn">Delete Post</button></a>
-        <a class="btn_a" href='/send-mail'><button class="btn admin-btn" action="/send-mail">Send Mail</button></a>
-        <a class="btn_a" href='/logout'><button class="btn admin-btn">Logout</button></a>
-    </div>
-{% endblock %}
