+++ application/templates/admin/create.html
@@ -36,7 +36,7 @@ function postPreview() {
 }
 </script>
 
-<form action="/create" method="POST" class="container" enctype="multipart/form-data">
+<form action="/admin/create" method="POST" class="container" enctype="multipart/form-data">
     <h1>Create New Post (#{{ post_id }})</h1>
     <i>If you are uploading images:
         <ol>
