+++ application/templates/admin/edit.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block main %}
-<form action="/edit/{{ post.id }}" method="POST" class="container" enctype="multipart/form-data">
+<form action="/admin/edit/{{ post.id }}" method="POST" class="container" enctype="multipart/form-data">
     <h1>Edit Post #{{ post_id }}</h1>
     <i>If you are uploading images:
         <ol>
