+++ application/templates/admin/login.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block main %}
-    <form action="/login" method="POST" class="container" id='login'>
+    <form action="/admin/login" method="POST" class="container" id='login'>
         {% if message %}
         <p style="color: red;">{{ message }}</p>
         {% endif %}
