+++ application/templates/admin/send-mail.html
@@ -16,11 +16,11 @@
     }
 
     function sendTest() {
-        document.email_form.action = "/send-test"
+        document.email_form.action = "/admin/send-test"
     }
 
     function sendMail() {
-        document.email_form.action = "/send-mail"
+        document.email_form.action = "/admin/send-mail"
     }
 </script>
